enhancement - add role authentication

+++ app/Http/Controllers/TransactionController.php
@@ -33,20 +33,22 @@ class TransactionController extends Controller
     {
         $data = $request->validate([
             'ticket_id' => 'required|integer',
-            'amount_of_ticket' => 'required|integer'
+            'amount_of_ticket' => 'required|integer',
+            'role' => 'required|in:ARTISTS,DEVS,USERS'
         ]);
 
-        $ticket = Ticket::find($data['ticket_id']);
-        $ticket_price = $data['amount_of_ticket'] * $ticket->price;
-        $tax = $ticket_price * (11/100);
-        $total_price = $ticket_price + $tax;
+        if($data['role'] === "ARTISTS" || $data['role'] === "DEVS"){
+            $ticket = Ticket::find($data['ticket_id']);
+            $ticket_price = $data['amount_of_ticket'] * $ticket->price;
+            $tax = $ticket_price * (11/100);
+            $total_price = $ticket_price + $tax;
             $transactions = Transaction::create([
                 'transaction_id' => uniqid(),
                 'amount_of_ticket' => $request->amount_of_ticket,
                 'ticket_price' => $ticket_price,
                 'total_price' => $total_price,
                 'concert_name' => $ticket->concert_name,
-                'concert_address' => $ticket->address,
+                'concert_address' => $ticket->concert_address,
                 'concert_date' => $ticket->concert_date,
                 'currency' => $ticket->currency
             ]);
@@ -55,7 +57,13 @@ class TransactionController extends Controller
                 'message' => "data successfully created",
                 'data' => $transactions
             ], 200);
-
+        } else {
+            return response()->json([
+                'status' => 405,
+                'message' => "permission denied",
+                'data' => "$request->role don't have permission to this method"
+            ], 405);
+        }
     }
 
     /**
@@ -93,37 +101,46 @@ class TransactionController extends Controller
     {
         $data = $request->validate([
             'ticket_id' => 'required|integer',
-            'amount_of_ticket' => 'required|integer'
-
+            'amount_of_ticket' => 'required|integer',
+            'role' => 'required|in:ARTISTS,DEVS,USERS'
         ]);
 
-        $transactions = Transaction::find($id);
-        $ticket = Ticket::find($data['ticket_id']);
-        $ticket_price = $data['amount_of_ticket'] * $ticket->price;
-        $tax = $ticket_price * (11/100);
-        $total_price = $ticket_price + $tax;
-        if($transactions){
-            $transactions->transaction_id = $transactions->transaction_id;
-            $transactions->total_price = $total_price;
-            $transactions->amount_of_ticket = $request->amount_of_ticket;
-            $transactions->ticket_price = $ticket_price;
-            $transactions->concert_name = $ticket->concert_name;
-            $transactions->concert_address = $ticket->address;
-            $transactions->concert_date = $ticket->concert_date;
-            $transactions->currency = $ticket->currency;
-            $transactions->save();
-            return response()->json([
-                'status' => 200,
-                'message' => 'data successfully updated',
-                'data' => $transactions
-            ], 200);
+        
+        if($data['role'] === "ARTISTS" || $data['role'] === "DEVS"){
+            $transactions = Transaction::find($id);
+            $ticket = Ticket::find($data['ticket_id']);
+            $ticket_price = $data['amount_of_ticket'] * $ticket->price;
+            $tax = $ticket_price * (11/100);
+            $total_price = $ticket_price + $tax;
+            if($transactions){
+                $transactions->transaction_id = $transactions->transaction_id;
+                $transactions->total_price = $total_price;
+                $transactions->amount_of_ticket = $request->amount_of_ticket;
+                $transactions->ticket_price = $ticket_price;
+                $transactions->concert_name = $ticket->concert_name;
+                $transactions->concert_address = $ticket->concert_address;
+                $transactions->concert_date = $ticket->concert_date;
+                $transactions->currency = $ticket->currency;
+                $transactions->save();
+                return response()->json([
+                    'status' => 200,
+                    'message' => 'data successfully updated',
+                    'data' => $transactions
+                ], 200);
+            } else {
+                return response()->json([
+                    'status' => 404,
+                    'message' => "transaction not found", 
+                    'data' => "transaction with $id not found",
+                ], 404);
+            };   
         } else {
             return response()->json([
-                'status' => 404,
-                'message' => "transaction not found", 
-                'data' => "transaction with $id not found",
-            ], 404);
-        };
+                'status' => 405,
+                'message' => "permission denied",
+                'data' => "$request->role don't have permission to this method"
+            ], 405);
+        }
     }
 
     /**
@@ -132,22 +149,36 @@ class TransactionController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy($id, Request $request)
     {
-        $transactions = Transaction::where('id', $id)->first();
-        if($transactions){
-            $transactions->delete();
-            return response()->json([
-                'status' => 200,
-                'message' => "data successfully deleted",
-                'data' => $transactions
-            ], 200);
+        $data = $request->validate([
+            'ticket_id' => 'required|integer',
+            'amount_of_ticket' => 'required|integer',
+            'role' => 'required|in:ARTISTS,DEVS,USERS'
+        ]);
+
+        if($data['role'] === "ARTISTS" || $data['role'] === "DEVS"){
+            $transactions = Transaction::where('id', $id)->first();
+            if($transactions){
+                $transactions->delete();
+                return response()->json([
+                    'status' => 200,
+                    'message' => "data successfully deleted",
+                    'data' => $transactions
+                ], 200);
+            } else {
+                return response()->json([
+                    'status' => 404,
+                    'message' => "transaction not found",
+                    'data' => "transaction with $id not found"
+                ], 404);
+            }   
         } else {
             return response()->json([
-                'status' => 404,
-                'message' => "transaction not found",
-                'data' => "transaction with $id not found"
-            ], 404);
+                'status' => 405,
+                'message' => "permission denied",
+                'data' => "$request->role don't have permission to this method"
+            ], 405);
         }
     }
 }
